refactor: Rename clients
