+++ rossum_api/__init__.py
@@ -1,14 +1,14 @@
 from __future__ import annotations
 
-from rossum_api.api_client import APIClientError
-from rossum_api.elis_api_client import ElisAPIClient, ExportFileFormats
-from rossum_api.elis_api_client_sync import ElisAPIClientSync
+from rossum_api.clients.external_async_client import AsyncRossumAPIClient, ExportFileFormats
+from rossum_api.clients.external_sync_client import SyncRossumAPIClient
+from rossum_api.clients.internal_async_client import APIClientError
 
 __version__ = "0.20.0"
 
 __all__ = (
     "APIClientError",
-    "ElisAPIClient",
-    "ElisAPIClientSync",
+    "AsyncRossumAPIClient",
+    "SyncRossumAPIClient",
     "ExportFileFormats",
 )
