+++ tests/test_api_client.py
@@ -19,7 +19,7 @@ import httpx
 import pytest
 import pytest_httpx
 
-from rossum_api.api_client import APIClient, APIClientError
+from rossum_api.clients.internal_async_client import APIClientError, InternalAsyncClient
 from rossum_api.domain_logic.resources import Resource
 
 WORKSPACES = [
@@ -150,7 +150,9 @@ def count_calls(func):
 @pytest.fixture
 def client():
     # Set retrying parameters to zero to keep tests fast
-    client = APIClient("username", "password", retry_backoff_factor=0, retry_max_jitter=0)
+    client = InternalAsyncClient(
+        "username", "password", retry_backoff_factor=0, retry_max_jitter=0
+    )
     client.token = FAKE_TOKEN
     return client
 
@@ -207,7 +209,7 @@ async def test_authenticate_is_retried(client, httpx_mock):
 @pytest.mark.asyncio
 async def test_init_token(httpx_mock):
     """Verifies that we can create client using token and call API without 'auth' call."""
-    client = APIClient(token=FAKE_TOKEN)
+    client = InternalAsyncClient(token=FAKE_TOKEN)
 
     httpx_mock.add_response(
         method="GET",
@@ -222,7 +224,7 @@ async def test_init_token(httpx_mock):
 @pytest.mark.asyncio
 async def test_reauth_no_credentials(httpx_mock):
     """Invalid token used but no credentials available for re-authentication. Raise 401."""
-    client = APIClient(token=FAKE_TOKEN)
+    client = InternalAsyncClient(token=FAKE_TOKEN)
 
     httpx_mock.add_response(
         method="GET",
@@ -239,7 +241,7 @@ async def test_reauth_no_credentials(httpx_mock):
 @pytest.mark.asyncio
 async def test_reauth_success(login_mock, httpx_mock):
     """Invalid token used, credentials available => reauthenticate in the background. Fetch user without raising any errors."""
-    client = APIClient("username", "password", token=FAKE_TOKEN)
+    client = InternalAsyncClient("username", "password", token=FAKE_TOKEN)
 
     httpx_mock.add_response(
         method="GET",
@@ -678,7 +680,7 @@ async def test_authenticate_generator_if_needed_token_expired(client, httpx_mock
 
 @pytest.mark.asyncio
 async def test_authenticate_if_needed_no_token(httpx_mock):
-    client = APIClient("username", "password")
+    client = InternalAsyncClient("username", "password")
     httpx_mock.add_response(
         method="GET",
         url="https://elis.rossum.ai/api/v1/workspaces/7694",
@@ -697,7 +699,7 @@ async def test_authenticate_if_needed_no_token(httpx_mock):
 
 @pytest.mark.asyncio
 async def test_authenticate_generator_if_needed_no_token(client, httpx_mock):
-    client = APIClient("username", "password")
+    client = InternalAsyncClient("username", "password")
     httpx_mock.add_response(
         method="GET",
         url="https://elis.rossum.ai/api/v1/queues/123/export?format=csv",
